+++ addOns/dev/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Changed
+- Sequence performance test to make it actually possible to test it using automation.
 
 ## [0.7.0] - 2024-10-07
 ### Added